modify base_code.c
